HYDRATOR-378 add checkpointing to data streams app
This involves adding a FileSet that will be used to write
checkpoints for all pipelines. The change also requires some
refactoring since all driver functions now need to be
serializable, due to how Spark streaming checkpoints work.
As such, moving all the anonymous classes in the DStreamCollection
class into their own classes. Also had to tweak how SparkCompute
plugins are handled so that the plugin is instantiated and
initialized every time instead of just once in the driver. This
is so that macro evaluation will be run each time a pipeline is
run, instead of just the very first time the pipeline is run and
then serialized.